several changes: - correct typings - move index as package.json main should be used - change webpack config to build a umd library
