--- index.ts
@@ -1,4 +1,0 @@
-/// <reference path="./lib/pigeon.d.ts"/>
-
-import Pigeon from './lib/pigeon.ts';
-export default new Pigeon();
