+++ lib/index.ts
@@ -1,0 +1,4 @@
+/// <reference path="./pigeon.d.ts"/>
+
+import Pigeon from './pigeon.ts';
+export default new Pigeon();
